// src/types/chat/job-posting/job-posting-chat-message-type.ts
import { FieldValue, Timestamp } from "firebase/firestore";
export enum JobPostingChatMessageTypeEnum {
  TEXT = "text",
  IMAGE = "image",
  FILE = "file",
  JOB_POSTING = "jobPosting",
  RESUME = "resume",
  SYSTEM = "system",
}

export interface MetaPathType {
  [key: string]: string;
}

export interface JobPostingChatMessageType {
  id: string;
  message: string;
  messageType: JobPostingChatMessageTypeEnum;
  metaPathList: MetaPathType[];
  senderId: string;

  createdAt: Timestamp | FieldValue;
  updatedAt: Timestamp | FieldValue;
}